feat(interface-types) Rename `lowering_lifting` module to `numbers`.
